Adjust decorator naming to reflect purpose
Some decorators and helpers naming didn't reflect the purpose of
the class properly.
